requirements.txt
langchain==0.3.7
langchain-google-genai==2.1.12
pydantic==2.11.9
pydantic-settings==2.10.1
pydantic_core==2.33.2
python-dotenv==1.1.1
streamlit==1.50.0